% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tvtdata}
\alias{tvtdata}
\title{Testing Data}
\format{
a dataframe with 8 rows and 5 columns:

\describe{
\item{stos}{Date. Starting date of the study.}
\item{treatna}{Date. Date of the time-varying treatment. Contains NA's for untreated patients.}
\item{eos}{Date. Date of the end of the study.}
\item{eos01}{Integer. State at the end of the study.}
\item{fct}{Character. A 2 level factor.}
}
}
\usage{
tvtdata
}
\description{
A simulated dataset
}
\keyword{datasets}
