% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tvtdata2}
\alias{tvtdata2}
\title{Testing Data 2}
\format{
a dataframe with 200 rows and 9 columns:

\describe{
\item{stos}{Date. Starting date of the study.}
\item{treatna}{Date. Date of the time-varying treatment. Contains NA's for untreated patients.}
\item{eos}{Date. Date of the end of the study.}
\item{eostreat}{Date. Date of the end of the period at risk of treatment.}
\item{eos01}{Integer. State at the end of the study.}
\item{fct_2lvl}{Character. A 2 level factor.}
\item{fct_4lvl}{Character. A 4 level factor.}
\item{xcov}{Integer. A covariate of numeric.}
}
}
\usage{
tvtdata2
}
\description{
A simulated dataset
}
\keyword{datasets}
